pyproject.toml
[project]
name = "mcp-server-jysd-cx-oracle"
version = "1.0.1"
description = "Model Context Protocol server to access oracle database using cx_Oracle"
authors = [
    { name = "hdcola", email = "gh@hdcola.org" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.5.0",
    "cx_Oracle>=8.3.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
mcp-server-jysd-cx-oracle = "mcp_server_jysd_cx_oracle:main"
dev = "mcp_server_jysd_cx_oracle:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
